requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
bcrypt==4.0.1
psycopg2-binary==2.9.9
sentry-sdk[flask]==1.38.0
python-dotenv==1.0.0